refactor: move record related components to record directory
