+++ app/lib/record/records_dao.dart
@@ -1,6 +1,6 @@
 import 'package:sqflite/sqflite.dart';
 
-import 'database_helper.dart';
+import '../database_helper.dart';
 import 'record.dart';
 
 class RecordsDao {
